requirements.txt
pandas>=1.5.0
numpy>=1.21.0
streamlit>=1.28.0
plotly>=5.15.0
statsmodels>=0.14.0
scikit-learn>=1.3.0
openai>=1.14.0
python-dotenv>=1.0.0
requests>=2.31.0
matplotlib>=3.7.0
seaborn>=0.12.0